+++ src/profile/games/chlcc/moviemenu.cpp
@@ -54,7 +54,7 @@ void Configure() {
   ButtonPromptSprite = EnsureGetMember<Sprite>("ButtonPromptSprite");
 
   auto drawType = Game::DrawComponentType::_from_integral_unchecked(
-      EnsureGetMember<int>("DrawType"));
+      EnsureGetMember<uint8_t>("DrawType"));
 
   UI::Menus[drawType].push_back(new UI::CHLCC::MovieMenu());
 }
