+++ src/profile/games/chlcc/sysmesbox.cpp
@@ -36,7 +36,7 @@ void Configure() {
   LoadingStarsPosition = EnsureGetMember<glm::vec2>("LoadingStarsPosition");
 
   auto drawType = Game::DrawComponentType::_from_integral_unchecked(
-      EnsureGetMember<int>("DrawType"));
+      EnsureGetMember<uint8_t>("DrawType"));
 
   UI::SysMesBoxPtr = new UI::CHLCC::SysMesBox();
   UI::Menus[drawType].push_back(UI::SysMesBoxPtr);
